✨automatically swap to using dms when >1 real user joins the uno game, keeping their hands private

+++ util/game.js
@@ -7,52 +7,95 @@ const {
 const send = require("./send");
 const showHand = require("./showHand");
 const delay = require("./delay");
+const msgAllPlayers = require("./msgAllPlayers");
 
-async function resetGame(bot, msg) {
-	msg.channel.unoPlayers = [];
-	msg.channel.unoRunning = false;
-	bot.unogame = null;
-	await bot.webhooks.uno.delete();
+async function resetGame(bot) {
+	bot.unogame.unoPlayers = [];
+	bot.unogame.unoRunning = false;
+	bot.unogame = {};
+	if (bot.webhooks.uno.name === "UnoBot") {
+		await bot.webhooks.uno.delete();
+	}
 	bot.webhooks = {};
 }
 
 async function nextTurn(bot, msg, players) {
+	console.log(players);
 	if (!bot.unogame) {
-		return null;
+		return players;
 	}
-	const member = msg.guild.members.cache.get(bot.unogame.currentPlayer.name);
+	const user = await bot.users.fetch(bot.unogame.currentPlayer.name);
 	const name = bot.unogame.discardedCard.toString();
 	const n2 = (name.includes("WILD_DRAW")) ? `${name.split(" ")[0]} WD4` : (name.includes("DRAW")) ? `${name.split(" ")[0]} DT` : name;
-	await send(bot.webhooks.uno, `You're up ${member} - Card: ${n2}`, {
-		files: [getCardImage(bot.unogame.discardedCard)],
-	});
-	if (players) {
-		players.forEach((p) => {
-			showHand(bot, msg, p);
+	if (!players.includes(bot.user.id)) {
+		bot.webhooks.uno = await bot.users.fetch(bot.unogame.currentPlayer.name);
+	}
+	if (!players.includes(bot.user.id)) {
+		const previousPlayer = {
+			id: null,
+		};
+		await msgAllPlayers(bot, players, previousPlayer, `${user} is up - Card: ${n2}`, {
+			files: [getCardImage(bot.unogame.discardedCard)],
 		});
 	} else {
-		showHand(bot, msg, bot.unogame.currentPlayer);
+		await send(bot.webhooks.uno, `${user} is up - Card: ${n2}`, {
+			files: [getCardImage(bot.unogame.discardedCard)],
+		});
 	}
+	// if (players) {
+	// 	players.forEach((p) => {
+	// 		showHand(bot, msg, p);
+	// 	});
+	// } else {
+	// 	showHand(bot, msg, bot.unogame.currentPlayer);
+	// }
+	showHand(bot, msg, bot.unogame.currentPlayer, players);
 	if (bot.unogame.currentPlayer.name === bot.user.id) {
 		await delay(2000);
-		return bot.user.id;
+		return null;
 		/* `doBotTurn(bot, msg);`
 			but then recursive requires,
 			so pass a value back up the stack instead */
 	}
-	return null;
+	return players;
+}
+
+async function sendWinMessage(bot, winner, score, players) {
+	console.log(players);
+	if (players.length === 2 && !players.includes(bot.user.id)) {
+		await send(bot.webhooks.uno, `${winner} wins! Score: ${score}`);
+	} else {
+		const previousPlayer = {
+			id: null,
+		};
+		await msgAllPlayers(bot, players, previousPlayer, `${winner} wins! Score: ${score}`);
+		// const fetchPlayers = players.map(p => bot.users.fetch(p));
+		// const users = await Promise.all(fetchPlayers);
+		// const sendUsers = users.map(u => send(u, `${winner} wins! Score: ${score}`));
+		// await Promise.all(sendUsers);
+
+		// for (const player of players) {
+		// 	bot.users.fetch(player).then((user) => {
+		// 		send(user, `${winner} wins! Score: ${score}`);
+		// 	});
+		// }
+	}
 }
 
 async function startGame(bot, msg, players) {
 	bot.unogame = new Game(players);
 	bot.unogame.newGame();
+	console.log(players);
 	bot.unogame.on("end", async (err, winner, score) => {
-		const memb = msg.guild.members.cache.get(winner.name); // winner.name === member.id
-		await send(bot.webhooks.uno, `${memb} wins! Score: ${score}`);
+		console.log(players);
+		const user = await bot.users.fetch(winner.name); // winner.name === member.id
+		await sendWinMessage(bot, user, score, players);
 		await resetGame(bot, msg);
 	});
-	msg.channel.unoRunning = true;
-	await send(bot.webhooks.uno, `${msg.author} has started Uno!`);
+	bot.unogame.unoRunning = true;
+	await send(bot.webhooks.uno, `${msg.author} has started Uno!${(players.includes(bot.user.id)) ? "" : " The game will be played in your DMs to keep your hand private!"}`);
+
+	await delay(5000);
 	const check = await nextTurn(bot, msg, players); // Pass the bot id all the way back up the stack
 	return check;
 }
@@ -63,10 +106,20 @@ async function beginning(bot, hook, msg, players) {
 	const collector = msg.channel.createMessageCollector(m => m.content.toLowerCase() === "join", {
 		time: 30000,
 	});
-	collector.on("collect", (m) => {
+	collector.on("collect", async (m) => {
 		if (!players.includes(m.author.id) && players.length < 10) {
-			send(hook, `${m.guild.members.cache.get(m.author.id).displayName} will play!`);
-			players.push(m.author.id);
+			let canEnter = true;
+			try {
+				await send(m.author, "You are entered to play Uno. The game will be run in DMs.");
+			} catch (e) {
+				console.log(e);
+				canEnter = false;
+				await send(hook, `${m.guild.members.cache.get(m.author.id)} Unable to send you a DM. Uno requires DMs.`);
+			}
+			if (canEnter) {
+				await send(hook, `${m.guild.members.cache.get(m.author.id).displayName} will play!`);
+				players.push(m.author.id);
+			}
 		}
 	});
 	await new Promise((resolve, reject) => { // This exists
